+++ src/widgets/gesture_detector.cpp
@@ -11,6 +11,7 @@ namespace systems::leal::campello_widgets
         ro->on_tap         = on_tap;
         ro->on_double_tap  = on_double_tap;
         ro->on_long_press  = on_long_press;
+        ro->on_pan_down    = on_pan_down;
         ro->on_pan_update  = on_pan_update;
         ro->on_pan_end     = on_pan_end;
         ro->on_scroll      = on_scroll;
@@ -23,6 +24,7 @@ namespace systems::leal::campello_widgets
         ro.on_tap          = on_tap;
         ro.on_double_tap   = on_double_tap;
         ro.on_long_press   = on_long_press;
+        ro.on_pan_down     = on_pan_down;
         ro.on_pan_update   = on_pan_update;
         ro.on_pan_end      = on_pan_end;
         ro.on_scroll       = on_scroll;